# swagger.yml
openapi: 3.0.0
info:
  title: Reservation API
  description: API for Reservation service
  version: 0.0.1

servers:
  - url: localhost:8000

paths:
  /:
    get:
      summary: Returns all existing reservations for all Restaurants
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReservationModel'
    post:
      summary: Creates a Reservation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReservationRequest'
      responses:
        '200':
          description: Successful creation of Reservation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationModel'
  /{id}:
    get:
      summary: Returns the reservation
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The Reservation ID
      responses:
        '200':
          description: Successful Retrival of reservation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationModel'
    delete:
      summary: Deletes the reservation
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The Reservation ID
      responses:
        '204':
          description: Sucessful reception of delete request
          content:
            application/json:
              schema:
                type: string

  /createMany:
    post:
      summary: Creates multiple Reservations
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/ReservationRequest'
      responses:
        '201':
          description: Succesful operation
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: The error message
                  invalid:
                    type: array
                    items:
                      $ref: '#/components/schemas/ReservationRequest'
                    description: The reservations that failed validation

  /deleteMany/{ids}:
    post:
      summary: Deletes multiple Reservations
      parameters:
        - in: path
          name: ids
          required: true
          schema:
            type: array
            items:
              type: string
              description: The reservation ID
      responses:
        '204':
          description: Successful reception of delete request
components:
  schemas:
    ReservationRequest:
      type: object
      properties:
        restaurant:
          type: string
          description: The Restaurant Name
        name:
          type: string
          description: The name on the Reservations
        timeSlot:
          type: string
          description: The time of the Reservation
        guests:
          type: integer
          description: The total number of guests, including children
        children:
          type: integer
          description: The number of children in the Reservation

    ReservationModel:
      type: object
      properties:
        _id:
          type: string
          description: The Reservation ID
        restaurant:
          type: string
          description: The Restaurant Name
        name:
          type: string
          description: The name on the Reservations
        timeSlot:
          type: string
          description: The time of the Reservation
        guests:
          type: integer
          description: The total number of guests, including children
        children:
          type: integer
          description: The number of children in the Reservation